Bugfix dealing with the way Cell component was handling width prop
An issue has persisted since the inception of the Cell component where
the first value received by the width prop of the Cell was being added
under a min-width of 1024px, the current desktop breakpoint. The problem with
this is that for screens smaller than this breakpoint, they would not receive
the proper width styling.

To fix this, the Cell component is now a mobile-first component, and the first
value received by the width prop is a base width. This base width is not added
under a media query.

    - Updated README
    - Added test to account for this
    - Fixed broken snapshots